--- src/test/java/runkoserver/controller/RunkoControllerTest.java
@@ -1,54 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package runkoserver.controller;
-
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-/**
- *
- * @author sieni
- */
-public class RunkoControllerTest {
-    
-    public RunkoControllerTest() {
-    }
-    
-    @BeforeClass
-    public static void setUpClass() {
-    }
-    
-    @AfterClass
-    public static void tearDownClass() {
-    }
-    
-    @Before
-    public void setUp() {
-    }
-    
-    @After
-    public void tearDown() {
-    }
-
-//    /**
-//     * Test of Hello method, of class RunkoController.
-//     */
-//    @Test
-//    public void testHello() {
-//        System.out.println("Hello");
-//        RunkoController instance = new RunkoController();
-//        String expResult = "";
-//        String result = instance.Hello();
-//        assertEquals(expResult, result);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-    
-}
